requirements.txt
appdirs==1.4.3
aspy.refactor-imports==2.0.0
atomicwrites==1.3.0
attrs==19.3.0
black==20.8b1
cached-property==1.5.1
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.0.3
distlib==0.3.0
filelock==3.0.12
identify==1.4.11
idna==2.10
importlib-metadata==1.5.0
isort==4.3.21
more-itertools==8.2.0
mypy-extensions==0.4.3
nodeenv==1.3.5
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
pre-commit==2.1.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
python-dotenv==0.15.0
PyYAML==5.3
regex==2020.2.20
requests==2.24.0
rope==0.16.0
seed-isort-config==2.1.0
six==1.14.0
spotipy==2.16.1
toml==0.10.2
tox==3.14.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.0.8
voluptuous==0.11.7
wcwidth==0.1.8
zipp==3.1.0
